执行器表属性调整，"顺序" 属性调整为整型，解决执行器数据较多时无法正确排序的问题；

+++ doc/db/tables_xxl_job.sql
@@ -78,7 +78,7 @@ CREATE TABLE `xxl_job_group` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `app_name` varchar(64) NOT NULL COMMENT '执行器AppName',
   `title` varchar(12) NOT NULL COMMENT '执行器名称',
-  `order` tinyint(4) NOT NULL DEFAULT '0' COMMENT '排序',
+  `order` int(11) NOT NULL DEFAULT '0' COMMENT '排序',
   `address_type` tinyint(4) NOT NULL DEFAULT '0' COMMENT '执行器地址类型：0=自动注册、1=手动录入',
   `address_list` varchar(512) DEFAULT NULL COMMENT '执行器地址列表，多地址逗号分隔',
   PRIMARY KEY (`id`)
